refactor: split Editor.tsx into smaller components and hooks
Break the 2300-line monolithic Editor component into focused modules:

- FormatBar.tsx: Toolbar with formatting buttons and table grid picker
- EditorTopBar.tsx: Top bar with save status, pin, search, export controls
- SearchHighlightExtension.ts: TipTap extension for search match highlighting
- useNoteSaving.ts: Save/load logic, source mode, auto-save debouncing
- useEditorExtensions.ts: TipTap editor setup with all extensions
- useEditorSearch.ts: Find-in-note with decorations and navigation
- usePopupManager.ts: Link, block math, and image popup management
- useCopyExport.ts: Copy and export handlers (markdown, plain text, HTML, PDF)
- useTableContextMenu.ts: Table right-click context menu via Tauri

Editor.tsx is now a ~530-line orchestrator that composes these hooks and
components, down from 2300 lines (77% reduction).
