+++ src/components/editor/SearchHighlightExtension.ts
@@ -1,0 +1,50 @@
+import { Extension } from "@tiptap/core";
+import { DecorationSet } from "@tiptap/pm/view";
+import { Plugin, PluginKey } from "@tiptap/pm/state";
+
+// Search highlight extension - adds yellow backgrounds to search matches
+export const searchHighlightPluginKey = new PluginKey("searchHighlight");
+
+interface SearchHighlightOptions {
+  matches: Array<{ from: number; to: number }>;
+  currentIndex: number;
+}
+
+export const SearchHighlight = Extension.create<SearchHighlightOptions>({
+  name: "searchHighlight",
+
+  addOptions() {
+    return {
+      matches: [],
+      currentIndex: 0,
+    };
+  },
+
+  addProseMirrorPlugins() {
+    return [
+      new Plugin({
+        key: searchHighlightPluginKey,
+        state: {
+          init: () => DecorationSet.empty,
+          apply: (tr, oldSet) => {
+            // Map decorations through document changes
+            const set = oldSet.map(tr.mapping, tr.doc);
+
+            // Check if we need to update decorations (from transaction meta)
+            const meta = tr.getMeta(searchHighlightPluginKey);
+            if (meta !== undefined) {
+              return meta.decorationSet;
+            }
+
+            return set;
+          },
+        },
+        props: {
+          decorations: (state) => {
+            return searchHighlightPluginKey.getState(state);
+          },
+        },
+      }),
+    ];
+  },
+});
