+++ src/components/editor/useEditorExtensions.ts
@@ -1,0 +1,210 @@
+import { type RefObject } from "react";
+import {
+  useEditor,
+  ReactNodeViewRenderer,
+  type Editor as TiptapEditor,
+} from "@tiptap/react";
+import StarterKit from "@tiptap/starter-kit";
+import Placeholder from "@tiptap/extension-placeholder";
+import Link from "@tiptap/extension-link";
+import Image from "@tiptap/extension-image";
+import TaskList from "@tiptap/extension-task-list";
+import TaskItem from "@tiptap/extension-task-item";
+import { TableKit } from "@tiptap/extension-table";
+import { Markdown } from "@tiptap/markdown";
+import CodeBlockLowlight from "@tiptap/extension-code-block-lowlight";
+import { lowlight } from "./lowlight";
+import { CodeBlockView } from "./CodeBlockView";
+import { Frontmatter } from "./Frontmatter";
+import { SearchHighlight } from "./SearchHighlightExtension";
+import { SlashCommand } from "./SlashCommand";
+import { Wikilink } from "./Wikilink";
+import { WikilinkSuggestion } from "./WikilinkSuggestion";
+import { ScratchBlockMath } from "./MathExtensions";
+import { invoke, convertFileSrc } from "@tauri-apps/api/core";
+import { join } from "@tauri-apps/api/path";
+import { toast } from "sonner";
+
+// Standard number-field shortcuts for KaTeX (shared between inline and block math)
+const katexMacros: Record<string, string> = {
+  "\\R": "\\mathbb{R}",
+  "\\N": "\\mathbb{N}",
+  "\\Z": "\\mathbb{Z}",
+  "\\Q": "\\mathbb{Q}",
+  "\\C": "\\mathbb{C}",
+};
+
+interface UseEditorExtensionsOptions {
+  isLoadingRef: RefObject<boolean>;
+  editorRef: RefObject<TiptapEditor | null>;
+  scheduleSave: () => void;
+  handleEditBlockMath: (pos: number) => void;
+  onSelectionUpdate?: () => void;
+}
+
+export function useEditorExtensions({
+  isLoadingRef,
+  editorRef,
+  scheduleSave,
+  handleEditBlockMath,
+  onSelectionUpdate,
+}: UseEditorExtensionsOptions): TiptapEditor | null {
+  const editor = useEditor({
+    extensions: [
+      StarterKit.configure({
+        heading: {
+          levels: [1, 2, 3, 4, 5, 6],
+        },
+        codeBlock: false,
+      }),
+      CodeBlockLowlight.extend({
+        addNodeView() {
+          return ReactNodeViewRenderer(CodeBlockView);
+        },
+      }).configure({
+        lowlight,
+        defaultLanguage: null,
+      }),
+      Placeholder.configure({
+        placeholder: "Start writing...",
+      }),
+      Link.configure({
+        openOnClick: false,
+        HTMLAttributes: {
+          class: "underline cursor-pointer",
+        },
+      }),
+      Image.configure({
+        inline: false,
+        allowBase64: false,
+      }),
+      TaskList,
+      TaskItem.configure({
+        nested: true,
+      }),
+      TableKit.configure({
+        table: {
+          resizable: false,
+          HTMLAttributes: {
+            class: "not-prose",
+          },
+        },
+      }),
+      Frontmatter,
+      Markdown.configure({}),
+      SearchHighlight.configure({
+        matches: [],
+        currentIndex: 0,
+      }),
+      SlashCommand,
+      Wikilink,
+      WikilinkSuggestion,
+      ScratchBlockMath.configure({
+        katexOptions: {
+          throwOnError: false,
+          displayMode: true,
+          macros: katexMacros,
+        },
+        onClick: (_node, pos) => {
+          handleEditBlockMath(pos);
+        },
+      }),
+    ],
+    editorProps: {
+      attributes: {
+        class:
+          "prose prose-lg dark:prose-invert max-w-3xl mx-auto focus:outline-none min-h-full px-6 pt-8 pb-24",
+      },
+      handleKeyDown: (_view, event) => {
+        if (event.key === "Tab") {
+          return false;
+        }
+        return false;
+      },
+      handlePaste: (_view, event) => {
+        const clipboardData = event.clipboardData;
+        if (!clipboardData) return false;
+
+        // Check for images first
+        const items = Array.from(clipboardData.items);
+        const imageItem = items.find((item) => item.type.startsWith("image/"));
+
+        if (imageItem) {
+          const blob = imageItem.getAsFile();
+          if (blob) {
+            const reader = new FileReader();
+            reader.onload = async () => {
+              const base64 = (reader.result as string).split(",")[1];
+
+              try {
+                const relativePath = await invoke<string>(
+                  "save_clipboard_image",
+                  { base64Data: base64 },
+                );
+                const notesFolder = await invoke<string>("get_notes_folder");
+                const absolutePath = await join(notesFolder, relativePath);
+                const assetUrl = convertFileSrc(absolutePath);
+
+                editorRef.current
+                  ?.chain()
+                  .focus()
+                  .setImage({ src: assetUrl })
+                  .run();
+              } catch (error) {
+                console.error("Failed to paste image:", error);
+                toast.error("Failed to paste image");
+              }
+            };
+            reader.onerror = () => {
+              console.error("Failed to read clipboard image:", reader.error);
+              toast.error("Failed to read clipboard image");
+            };
+            reader.readAsDataURL(blob);
+            return true;
+          }
+        }
+
+        // Handle markdown text paste
+        const text = clipboardData.getData("text/plain");
+        if (!text) return false;
+
+        const markdownPatterns =
+          /^#{1,6}\s|^\s*[-*+]\s|^\s*\d+\.\s|^\s*>\s|```|^\s*\[.*\]\(.*\)|^\s*!\[|\*\*.*\*\*|__.*__|~~.*~~|^\s*[-*_]{3,}\s*$|^\|.+\||\$\$[\s\S]+?\$\$/m;
+        if (!markdownPatterns.test(text)) {
+          return false;
+        }
+
+        const currentEditor = editorRef.current;
+        if (!currentEditor) return false;
+
+        const manager = currentEditor.storage.markdown?.manager;
+        if (manager && typeof manager.parse === "function") {
+          try {
+            const parsed = manager.parse(text);
+            if (parsed) {
+              currentEditor.commands.insertContent(parsed);
+              return true;
+            }
+          } catch {
+            // Fall back to default paste behavior
+          }
+        }
+
+        return false;
+      },
+    },
+    onCreate: ({ editor: editorInstance }) => {
+      editorRef.current = editorInstance;
+    },
+    onUpdate: () => {
+      if (isLoadingRef.current) return;
+      scheduleSave();
+    },
+    onSelectionUpdate: () => {
+      onSelectionUpdate?.();
+    },
+    immediatelyRender: false,
+  });
+
+  return editor;
+}
