+++ src/components/editor/usePopupManager.ts
@@ -1,0 +1,483 @@
+import { useEffect, useRef, useCallback, type RefObject } from "react";
+import {
+  ReactRenderer,
+  type Editor as TiptapEditor,
+} from "@tiptap/react";
+import { NodeSelection, TextSelection } from "@tiptap/pm/state";
+import tippy, { type Instance as TippyInstance } from "tippy.js";
+import { open as openDialog } from "@tauri-apps/plugin-dialog";
+import { invoke, convertFileSrc } from "@tauri-apps/api/core";
+import { join } from "@tauri-apps/api/path";
+import { toast } from "sonner";
+import { BlockMathEditor } from "./BlockMathEditor";
+import { LinkEditor } from "./LinkEditor";
+import { normalizeBlockMath } from "./MathExtensions";
+
+interface UsePopupManagerOptions {
+  editorRef: RefObject<TiptapEditor | null>;
+}
+
+interface UsePopupManagerReturn {
+  handleAddLink: () => void;
+  handleAddBlockMath: () => void;
+  handleEditBlockMath: (pos: number) => void;
+  handleAddImage: () => Promise<void>;
+}
+
+export function usePopupManager({
+  editorRef,
+}: UsePopupManagerOptions): UsePopupManagerReturn {
+  const linkPopupRef = useRef<TippyInstance | null>(null);
+  const blockMathPopupRef = useRef<TippyInstance | null>(null);
+
+  const closeBlockMathPopup = useCallback(() => {
+    if (blockMathPopupRef.current) {
+      blockMathPopupRef.current.destroy();
+      blockMathPopupRef.current = null;
+    }
+  }, []);
+
+  const handleEditBlockMath = useCallback(
+    (pos: number) => {
+      const currentEditor = editorRef.current;
+      if (!currentEditor) return;
+
+      if (linkPopupRef.current) {
+        linkPopupRef.current.destroy();
+        linkPopupRef.current = null;
+      }
+      closeBlockMathPopup();
+
+      const node = currentEditor.state.doc.nodeAt(pos);
+      if (!node || node.type.name !== "blockMath") {
+        return;
+      }
+
+      const virtualElement = {
+        getBoundingClientRect: () => {
+          const nodeDom = currentEditor.view.nodeDOM(pos);
+          if (nodeDom instanceof HTMLElement) {
+            return nodeDom.getBoundingClientRect();
+          }
+
+          const start = currentEditor.view.coordsAtPos(pos);
+          const end = currentEditor.view.coordsAtPos(pos + node.nodeSize);
+          const left = Math.min(start.left, end.left);
+          const top = Math.min(start.top, end.top);
+          const right = Math.max(start.right, end.right);
+          const bottom = Math.max(start.bottom, end.bottom);
+
+          return {
+            width: Math.max(2, right - left),
+            height: Math.max(20, bottom - top),
+            top,
+            left,
+            right,
+            bottom,
+            x: left,
+            y: top,
+            toJSON: () => ({}),
+          } as DOMRect;
+        },
+      };
+
+      const component = new ReactRenderer(BlockMathEditor, {
+        props: {
+          initialLatex: String(node.attrs.latex ?? ""),
+          onSubmit: (latex: string) => {
+            const trimmed = latex.trim();
+            if (!trimmed) {
+              toast.error("Please enter a formula.");
+              return;
+            }
+            currentEditor
+              .chain()
+              .focus()
+              .updateBlockMath({ pos, latex: trimmed })
+              .setTextSelection(pos + node.nodeSize)
+              .run();
+            closeBlockMathPopup();
+          },
+          onCancel: () => {
+            currentEditor
+              .chain()
+              .focus()
+              .setTextSelection(pos + node.nodeSize)
+              .run();
+            closeBlockMathPopup();
+          },
+        },
+        editor: currentEditor,
+      });
+
+      blockMathPopupRef.current = tippy(document.body, {
+        getReferenceClientRect: () =>
+          virtualElement.getBoundingClientRect() as DOMRect,
+        appendTo: () => document.body,
+        content: component.element,
+        showOnCreate: true,
+        interactive: true,
+        trigger: "manual",
+        placement: "bottom-start",
+        offset: [0, 8],
+        onDestroy: () => {
+          component.destroy();
+        },
+      });
+    },
+    [closeBlockMathPopup],
+  );
+
+  const handleAddBlockMath = useCallback(() => {
+    const currentEditor = editorRef.current;
+    if (!currentEditor) return;
+
+    closeBlockMathPopup();
+    if (linkPopupRef.current) {
+      linkPopupRef.current.destroy();
+      linkPopupRef.current = null;
+    }
+    const { selection, doc } = currentEditor.state;
+    const { from, to, empty, $from } = selection;
+
+    if (
+      selection instanceof NodeSelection &&
+      selection.node.type.name === "blockMath"
+    ) {
+      handleEditBlockMath(from);
+      return;
+    }
+
+    if (!empty) {
+      const selectedNode = doc.nodeAt(from);
+      if (
+        selectedNode?.type.name === "blockMath" &&
+        from + selectedNode.nodeSize === to
+      ) {
+        handleEditBlockMath(from);
+        return;
+      }
+    }
+
+    if (empty) {
+      const nodeBefore = $from.nodeBefore;
+      if (nodeBefore?.type.name === "blockMath") {
+        handleEditBlockMath(from - nodeBefore.nodeSize);
+        return;
+      }
+      const nodeAfter = $from.nodeAfter;
+      if (nodeAfter?.type.name === "blockMath") {
+        handleEditBlockMath(from);
+        return;
+      }
+    }
+
+    const selectedText = empty ? "" : doc.textBetween(from, to, "\n");
+    const initialLatex = normalizeBlockMath(selectedText);
+    const targetRange = { from, to };
+    const hasSelection = from !== to;
+
+    const virtualElement = {
+      getBoundingClientRect: () => {
+        if (hasSelection) {
+          const startPos = currentEditor.view.domAtPos(from);
+          const endPos = currentEditor.view.domAtPos(to);
+
+          if (startPos && endPos) {
+            try {
+              const range = document.createRange();
+              range.setStart(startPos.node, startPos.offset);
+              range.setEnd(endPos.node, endPos.offset);
+              return range.getBoundingClientRect();
+            } catch (error) {
+              console.error("Block math range creation failed:", error);
+            }
+          }
+        }
+
+        const coords = currentEditor.view.coordsAtPos(from);
+        return {
+          width: 2,
+          height: 20,
+          top: coords.top,
+          left: coords.left,
+          right: coords.right,
+          bottom: coords.bottom,
+          x: coords.left,
+          y: coords.top,
+          toJSON: () => ({}),
+        } as DOMRect;
+      },
+    };
+
+    const component = new ReactRenderer(BlockMathEditor, {
+      props: {
+        initialLatex,
+        onSubmit: (latex: string) => {
+          const normalizedLatex = latex.trim();
+          if (!normalizedLatex) {
+            toast.error("Please enter a formula.");
+            return;
+          }
+
+          const inserted = currentEditor
+            .chain()
+            .focus()
+            .insertContentAt(targetRange, {
+              type: "blockMath",
+              attrs: { latex: normalizedLatex },
+            })
+            .command(({ state, tr, dispatch }) => {
+              if (!dispatch) return true;
+
+              const { $to } = tr.selection;
+              if ($to.nodeAfter?.isTextblock) {
+                tr.setSelection(TextSelection.create(tr.doc, $to.pos + 1));
+                tr.scrollIntoView();
+                return true;
+              }
+
+              const paragraphType =
+                state.schema.nodes.paragraph ??
+                $to.parent.type.contentMatch.defaultType;
+              const paragraphNode = paragraphType?.create();
+              const insertPos = $to.nodeAfter ? $to.pos : $to.end();
+
+              if (paragraphNode) {
+                const $insertPos = tr.doc.resolve(insertPos);
+                if (
+                  $insertPos.parent.canReplaceWith(
+                    $insertPos.index(),
+                    $insertPos.index(),
+                    paragraphNode.type,
+                  )
+                ) {
+                  tr.insert(insertPos, paragraphNode);
+                  tr.setSelection(TextSelection.create(tr.doc, insertPos + 1));
+                  tr.scrollIntoView();
+                  return true;
+                }
+              }
+
+              tr.scrollIntoView();
+              return true;
+            })
+            .run();
+
+          if (inserted) {
+            closeBlockMathPopup();
+          }
+        },
+        onCancel: () => {
+          currentEditor.commands.focus();
+          closeBlockMathPopup();
+        },
+      },
+      editor: currentEditor,
+    });
+
+    blockMathPopupRef.current = tippy(document.body, {
+      getReferenceClientRect: () =>
+        virtualElement.getBoundingClientRect() as DOMRect,
+      appendTo: () => document.body,
+      content: component.element,
+      showOnCreate: true,
+      interactive: true,
+      trigger: "manual",
+      placement: "bottom-start",
+      offset: [0, 8],
+      onDestroy: () => {
+        component.destroy();
+      },
+    });
+  }, [closeBlockMathPopup, handleEditBlockMath]);
+
+  // Link handler - show inline popup at cursor position
+  const handleAddLink = useCallback(() => {
+    const editor = editorRef.current;
+    if (!editor) return;
+
+    closeBlockMathPopup();
+
+    if (linkPopupRef.current) {
+      linkPopupRef.current.destroy();
+      linkPopupRef.current = null;
+    }
+
+    const existingUrl = editor.getAttributes("link").href || "";
+    const { from, to } = editor.state.selection;
+    const hasSelection = from !== to;
+
+    const virtualElement = {
+      getBoundingClientRect: () => {
+        if (hasSelection) {
+          const startPos = editor.view.domAtPos(from);
+          const endPos = editor.view.domAtPos(to);
+
+          if (startPos && endPos) {
+            try {
+              const range = document.createRange();
+              range.setStart(startPos.node, startPos.offset);
+              range.setEnd(endPos.node, endPos.offset);
+              return range.getBoundingClientRect();
+            } catch (e) {
+              console.error("Range creation failed:", e);
+            }
+          }
+        }
+
+        const coords = editor.view.coordsAtPos(from);
+        return {
+          width: 2,
+          height: 20,
+          top: coords.top,
+          left: coords.left,
+          right: coords.right,
+          bottom: coords.bottom,
+          x: coords.left,
+          y: coords.top,
+          toJSON: () => ({}),
+        } as DOMRect;
+      },
+    };
+
+    const component = new ReactRenderer(LinkEditor, {
+      props: {
+        initialUrl: existingUrl,
+        initialText: hasSelection || existingUrl ? undefined : "",
+        onSubmit: (url: string, text?: string) => {
+          if (url.trim()) {
+            if (text !== undefined) {
+              if (text.trim()) {
+                editor
+                  .chain()
+                  .focus()
+                  .insertContent({
+                    type: "text",
+                    text: text.trim(),
+                    marks: [{ type: "link", attrs: { href: url.trim() } }],
+                  })
+                  .run();
+              }
+            } else {
+              editor
+                .chain()
+                .focus()
+                .extendMarkRange("link")
+                .setLink({ href: url.trim() })
+                .run();
+            }
+          } else {
+            editor.chain().focus().extendMarkRange("link").unsetLink().run();
+          }
+          linkPopupRef.current?.destroy();
+          linkPopupRef.current = null;
+        },
+        onRemove: () => {
+          editor.chain().focus().extendMarkRange("link").unsetLink().run();
+          linkPopupRef.current?.destroy();
+          linkPopupRef.current = null;
+        },
+        onCancel: () => {
+          editor.commands.focus();
+          linkPopupRef.current?.destroy();
+          linkPopupRef.current = null;
+        },
+      },
+      editor,
+    });
+
+    linkPopupRef.current = tippy(document.body, {
+      getReferenceClientRect: () =>
+        virtualElement.getBoundingClientRect() as DOMRect,
+      appendTo: () => document.body,
+      content: component.element,
+      showOnCreate: true,
+      interactive: true,
+      trigger: "manual",
+      placement: "bottom-start",
+      offset: [0, 8],
+      onDestroy: () => {
+        component.destroy();
+      },
+    });
+  }, [closeBlockMathPopup]);
+
+  // Image handler
+  const handleAddImage = useCallback(async () => {
+    const editor = editorRef.current;
+    if (!editor) return;
+    const selected = await openDialog({
+      multiple: false,
+      filters: [
+        {
+          name: "Images",
+          extensions: ["png", "jpg", "jpeg", "gif", "webp", "svg"],
+        },
+      ],
+    });
+    if (selected) {
+      try {
+        const relativePath = await invoke<string>("copy_image_to_assets", {
+          sourcePath: selected as string,
+        });
+        const notesFolder = await invoke<string>("get_notes_folder");
+        const absolutePath = await join(notesFolder, relativePath);
+        const assetUrl = convertFileSrc(absolutePath);
+        editor.chain().focus().setImage({ src: assetUrl }).run();
+      } catch (error) {
+        console.error("Failed to add image:", error);
+      }
+    }
+  }, []);
+
+  // Listen for slash command image insertion
+  useEffect(() => {
+    const handler = () => handleAddImage();
+    window.addEventListener("slash-command-image", handler);
+    return () => window.removeEventListener("slash-command-image", handler);
+  }, [handleAddImage]);
+
+  // Listen for slash command block math insertion
+  useEffect(() => {
+    const handler = () => handleAddBlockMath();
+    window.addEventListener("slash-command-block-math", handler);
+    return () =>
+      window.removeEventListener("slash-command-block-math", handler);
+  }, [handleAddBlockMath]);
+
+  // Keyboard shortcut for Cmd+K to add link (only when editor is focused)
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if ((e.metaKey || e.ctrlKey) && e.key === "k") {
+        const target = e.target as HTMLElement;
+        const isInEditor = target.closest(".ProseMirror");
+        if (isInEditor && editorRef.current) {
+          e.preventDefault();
+          handleAddLink();
+        }
+      }
+    };
+    document.addEventListener("keydown", handleKeyDown);
+    return () => document.removeEventListener("keydown", handleKeyDown);
+  }, [handleAddLink]);
+
+  // Cleanup popups on unmount
+  useEffect(() => {
+    return () => {
+      if (linkPopupRef.current) {
+        linkPopupRef.current.destroy();
+      }
+      if (blockMathPopupRef.current) {
+        blockMathPopupRef.current.destroy();
+      }
+    };
+  }, []);
+
+  return {
+    handleAddLink,
+    handleAddBlockMath,
+    handleEditBlockMath,
+    handleAddImage,
+  };
+}
